refactor: reducing the size of variables in struct

+++ contracts/interfaces/ITreasury.sol
@@ -7,12 +7,12 @@ interface ITreasury {
     /// @param tokenAmount Token amount to get OLA for.
     /// @param token Token address.
     /// @param olaMintAmount Amount of OLA token issued.
-    function depositTokenForOLAS(uint256 tokenAmount, address token, uint256 olaMintAmount) external;
+    function depositTokenForOLAS(uint224 tokenAmount, address token, uint96 olaMintAmount) external;
 
     /// @dev Deposits ETH from protocol-owned service.
     /// @param serviceIds Set of service Ids.
     /// @param amounts Correspondent set of amounts.
-    function depositETHFromServices(uint256[] memory serviceIds, uint256[] memory amounts) external payable;
+    function depositETHFromServices(uint32[] memory serviceIds, uint96[] memory amounts) external payable;
 
     /// @dev Gets information about token being enabled.
     /// @param token Token address.
